Completed this stage of the project
Really not happy with how long it took me to do
this bit.
The code is really scrappy and doesn't look nice
but it functions and that is the main thing.

Had some real issues with being able to address
things within a function and it was causing all
sorts of problems.